requirements.txt
streamlit
requests
beautifulsoup4
pdfplumber
python-docx
python-dotenv
sentence-transformers
faiss-cpu
numpy
google-genai
openai
groq
